LPS-136630 Moving to different package
